using finite field generator from sst merlin topology modules
